services/telemetry-ingestion-api/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
google-cloud-pubsub==2.18.4
google-cloud-logging==3.6.0
pydantic==2.10.6